ENH isolated optimizers in classes
